+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## 0.2.4 (Jun 23, 2021)
+## 0.3.0 (Jun 28, 2021)
+
+* integrates with HeroVau.lt module
+* map some names in pathbuilder to names within foundry
+* added a blacklist of feats/abilities to prevent them from throwing an error, as they're added by other means (eg: class/heritage/ancestry related)
+* added some other ways to search for item/feat/ability names to get better matching
+
+## 0.2.4 (Jun 28, 2021)
 
 * initial release under me
 * fixed spell import feature